Some more features + rm unused classes
